created yelpUtility, cleaned up html-routes.js
